Added Filled Cauldron
Added tile entity (to hold the liquid information) and renderer (to show the liquid level), so the cauldron in this state cannot be moved by pistons.
Got cauldron filling with potions/removal with glass bottles implemented.  When the filled cauldron empties out entirely, it will revert back to a standard cauldron.
The cauldron holds three potion's worth (similar to normal cauldron functionality) except it is measured from 0 to 15.
Standing in the cauldron doesn't do much aside from extinguishing fires while using up a bit of the liquid.  It will be used to apply the effect(s) from the liquid, consume a similar amount of liquid should the effects be applied.
Started code for the Imbuing system, whereby an ingredient can be tossed into a filled cauldron with the required effect, amplifier and fluid level present to get a new item.
Added two new items, Bottle of Fire and Bottle of Soul Fire, made by using a glass bottle on the respective fire blocks, with a chance of extinguishing the fire.  These will be used as ingredients in the future.