Updated brewing station model to v2.
Made an actual model for the brewing station.  Needs texture work.  Needs a TileEntityRenderer added for rendering the fluid internally.

Removed the use of BasePotion in the potion fluid system.  It now just uses Potion to identify base potions, as the normal potion system does.  It was just an annoying step to take anyways.

Fixed the Bottle of Sunlight and Bottle of Void Essence sprites due to incorrect filenames.

Added imbuing recipe template to patchouli book.

Reverted filled cauldrons to drop just the normal empty cauldron, to keep inline with the normal cauldron handling when filled with water.